% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fastq.R
\name{get_fastq}
\alias{get_fastq}
\title{Download fastq files}
\usage{
get_fastq(srr_id, library_layout = c("SINGLE", "PAIRED"),
  get_sra_file = FALSE, sra_files_dir = NULL, n_thread, destdir)
}
\arguments{
\item{srr_id}{SRA run accession ID.}

\item{library_layout}{layout of the library used. Either 
\code{'SINGLE'} or \code{'PAIRED'}.}

\item{get_sra_file}{logical, whether to download SRA file
first and get fastq files afterwards or directly download
fastq files.}

\item{sra_files_dir}{directory where SRA files are saved.
If you use \code{get_sra_file=FALSE} then \code{sra_files_dir=NULL}.}

\item{n_thread}{number of cores to use.}

\item{destdir}{directory where all the results will be saved.}
}
\value{
A single fastq file will be generated for SINGLE end
reads and two files for PAIRED end reads.
}
\description{
\code{get_fastq} downloads fastq files using SRA toolkit. 
We recommend using Aspera for fast downloading. You need to 
install Aspera(\url{http://www.asperasoft.com/}) for using 
\code{ascp} option.
}
\examples{
srr_id="SRR6324192"
\dontrun{
get_fastq(srr_id=srr_id, library_layout="SINGLE",
get_sra_file=FALSE, 
sra_files_dir=NULL, n_thread=2,
destdir="/mnt/raid/test")
}

}
