% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.R
\name{get_metadata}
\alias{get_metadata}
\title{Download metadata from GEO and SRA}
\usage{
get_metadata(geo_series_acc)
}
\arguments{
\item{geo_series_acc}{GEO series accession ID.}
}
\value{
a list of GEO and SRA metadata.
}
\description{
Download metadata from GEO and SRA
}
\examples{
geo_series_acc="GSE107363"
\dontrun{
get_metadata(geo_series_acc=geo_series_acc)
}

}
